requirements.txt
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.24.0
ollama>=0.1.8
duckduckgo-search>=6.0.0
pyyaml>=6.0
scipy>=1.11.0
pynput>=1.7.6
